requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via omnixy (pyproject.toml)
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.22.1
    # via omnixy (pyproject.toml)
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   anthropic
async-timeout==5.0.1
    # via aiohttp
attrs==25.1.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cohere==5.13.12
    # via omnixy (pyproject.toml)
cryptography==44.0.1
    # via omnixy (pyproject.toml)
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
fastavro==1.10.0
    # via cohere
filelock==3.17.0
    # via huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via huggingface-hub
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.24.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.161.0
    # via google-generativeai
google-auth==2.38.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.4
    # via omnixy (pyproject.toml)
googleapis-common-protos==1.68.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   cohere
    #   openai
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.29.1
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.8.2
    # via openai
loguru==0.7.3
    # via omnixy (pyproject.toml)
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
openai==1.64.0
    # via omnixy (pyproject.toml)
packaging==24.2
    # via huggingface-hub
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   omnixy (pyproject.toml)
    #   cohere
    #   google-generativeai
    #   openai
pydantic-core==2.27.2
    # via
    #   cohere
    #   pydantic
pyparsing==3.2.1
    # via httplib2
python-dotenv==1.0.1
    # via omnixy (pyproject.toml)
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.3
    # via
    #   omnixy (pyproject.toml)
    #   cohere
    #   google-api-core
    #   huggingface-hub
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tokenizers==0.21.0
    # via cohere
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
types-requests==2.32.0.20241016
    # via cohere
typing-extensions==4.12.2
    # via
    #   anyio
    #   cohere
    #   google-generativeai
    #   huggingface-hub
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
uv==0.6.3
    # via omnixy (pyproject.toml)
yarl==1.18.3
    # via aiohttp
